requirements.txt
torch
numpy
polars
transformers
PyYAML
tables
torchvision
tensorboard
scikit-learn
pandas